Added time received timestamps
There are still known issues. First, though this could be
unrelated, the graphs sometimes stop plotting new data for a brief
period of time, though they don't freeze. During this time, the
timestamps are still updating correctly. The second issue is that,
when running the engineering dashboard separately from the driver
dashboard, the timestamps are not always accurate. Sometimes, the
timestamps will jump back in time for one or two datasets.