Re-open socket after lost connection and track lost connections
If the client errors out, it will attempt to reopen the socket at
one second intervals. Also, solar_car_connection is now an array,
so lost connections between the solar car and engineering dashboard
can be tracked in the graphs. Right now, the most recent value of
solar_car_connection is set to false whenever connection is lost,
and true is pushed to the front of the array whenever new data is
unpacked. Because of this, only the element of solar_car_connection
that corresponds to the last packet received before connection was
lost will be false.