Close to actual implementation
Retrieves the most recent timestamp table name, as well as the
first timestamp in the table. Then, retrieves new rows from that
table. There is still an issue with duplicate rows being retrieved.
I believe this is because rows are being requested at fixed
intervals, with no regard to whether or not the previous request
for rows has finished.