Update FaultsView and images
Update existing faults and add new faults to FaultsView. Also, the
images used by this component have been updated to reflect the changes
to FaultsView. Additionally, this commit introduces a bug fix to the
_checkBooleanData() function. Since boolean nominal ranges are 0-1 in
the data format, strict comparisons between the nominal and actual
values don't work correctly, so the comparison in the function has to
allow type coercion.